requirements.txt
wheel
pyaudio
speechrecognition
tts
gtts
playsound
pygobject
mecab-python3
unidic-lite
inflect
fuzzywuzzy
python-levenshtein
flask
opencv-python
face-recognition
pyopengl
glfw
noise
googletrans
openai
elevenlabs
selenium
lockfile